fix: modified the packaging and updated the Java version.
feat: created a new release to validate, the build workflow, phase 1

feat: added the new get version from pom action

fix: changed the approach to get the version out of the pom file, and evaluate the create release

fix: the workflow is invalid, fixing the syntax

feat: added the upload of binaries to the created release, so that they can land in the public space

chore: bump the release to test the new code in the workflow.

fix: correction in the path.

chore: updated the version in the pom so that we can test the workflow with the new fix.

chore: updated the path, we cannot reference a folder apparently, and bumpped the version in the pom to test

chore:tried another time to see if we can see this path

chore: added the debug information that we need

chore: fixed the debug information and bumped the build again

chore: list the contents of the folder staging to prepare the next steps

fix: added the correct path and test the jar upload

chore, tried a different synthax in the path

fix: there was a typo in the path, fixing it to make sure we get the correct file name

feat: create a package with both jar files name it with the version being built.

feat: improved the naming

feat: adding more logging to assess the path of the jar files after compression

feat: Upload the newly generated package that contains the two jar files inside

feat: started the integration of the SonarQube Step

chore: bumpped the release to start testing the new actions in the pipeline

feat: added the metrics and prometheus endpoints to the standard Springboot profile.

feat: improved the analysis

feat: added the needed details regarding sonarqube into the pom, this is now Swiss Post specific.

fix: make sure the needed endpoints are all visible, metrics and health