feature: create a release
- Add the workflow to build the Mag
- updated the branches that can trigger the workflow
- disable code coverage
- disable the depenecy graph
- add a unique name to the artifacts uploaded
fix: modified the packaging and updated the Java version.

feat: created a new release to validate, the build workflow, phase 1

feat: added the new get version from pom action

fix: changed the approach to get the version out of the pom file, and evaluate the create release

fix: the workflow is invalid, fixing the syntax

feat: added the upload of binaries to the created release, so that they can land in the public space

chore: bump the release to test the new code in the workflow.

fix: correction in the path.

chore: updated the version in the pom so that we can test the workflow with the new fix.

chore: updated the path, we cannot reference a folder apparently, and bumpped the version in the pom to test

chore:tried another time to see if we can see this path

chore: added the debug information that we need

chore: fixed the debug information and bumped the build again

chore: list the contents of the folder staging to prepare the next steps

fix: added the correct path and test the jar upload

chore, tried a different synthax in the path

fix: there was a typo in the path, fixing it to make sure we get the correct file name

feat: create a package with both jar files name it with the version being built.

feat: improved the naming

feat: adding more logging to assess the path of the jar files after compression

feat: Upload the newly generated package that contains the two jar files inside

feat: started the integration of the SonarQube Step

chore: bumpped the release to start testing the new actions in the pipeline

feat: added the metrics and prometheus endpoints to the standard Springboot profile.

feat: improved the analysis

feat: added the needed details regarding sonarqube into the pom, this is now Swiss Post specific.

fix: make sure the needed endpoints are all visible, metrics and health

fix: cleanup and version tagging

- modify the created release
- update other actions to reduce the warnings
- test the new tag action and the new create release action
- bump the version
- changed the tag creation
chore: generate build number
chore: bummped the release to run the changes made on Sonar

feat: add capability to manually run the workflow

feat: add sonar to this branch and bumpped the version

chore: use java 17
feat: send the log to the console
feature: whitelist the health checks endpoints in the Springboot Firewall
feature: manage logging and traceId propagation

feat: added more one log source.

feat: add the recommend post log format

feat: added traceid and spanid to the mag

feat: Started to add the new traceid propagation

feat: deleted a file

fix: SleuthConfiguration is fixed, removed the not needed custom filters, add debug TraceHandler to understand if the http headers are arriving

fix: removed zipkin from the configuration to avoid the errors due to lack of connectivity

feat: added a custom propagation, that fixes both B3 and W3C headers for traceid

chore: bump version to build image

fix: got rid of the system.out instructions

feat: enhanced the mag logging to avoid aggregation of records in Splunk

fix: convert idents to spaces

fix bump versio fix idents

fix: try to override IPF using an alternative way, added more log to prove that we initialize correctly the log.

chore: bump the version to make sure the build is working

fix, fixed runtime errors on the Iti67 route. bumped the version and added extra debug information for the new docker image migration

fix: tried to read from the environment variable

chore: added more log and tweaked the idp code to make sure our dynamic setting is taken into account

chore: added more log to understand how the metada is behaving in the new docker image

fix: changed the names of the variables to align everything

chore: added more debug information to narrow down how the file is being read

chore: tried to clean the extra character

fix: removed code no longer needed